test: move thread into realtime class even on EVENT__DISABLE_THREAD_SUPPORT
